Xcode project updated with mysql build. Xcode is configured to assume mysql lib and headers are installed under /usr/local/mysql. SQLite and PostgreSQL are part of Lion and no special configuration is needed for them.

+++ src/db/mysql/MysqlResultSet.c
@@ -210,7 +210,7 @@ const void *MysqlResultSet_getBlob(T R, int columnIndex, int *size) {
         if (R->columns[i].is_null) 
                 return NULL;
         ensureCapacity(R, i);
-        *size = R->columns[i].real_length;
+        *size = (int)R->columns[i].real_length;
         return R->columns[i].buffer;
 }
